requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pymongo==4.6.0
pydantic==2.11.9
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
sshtunnel==0.4.0
paramiko==3.4.0
